requirements.txt
annotated-types==0.7.0
anyio==4.12.1
black==24.8.0
certifi==2026.1.4
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
distro==1.9.0
fonttools==4.61.1
flake8==7.1.0
google-generativeai==0.8.5
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
joblib==1.5.3
kiwisolver==1.4.9
matplotlib==3.10.8
nltk==3.9.2
numpy==2.4.2
openai==2.18.0
packaging==26.0
pillow==12.1.0
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.2
python-dateutil==2.9.0.post0
regex==2026.1.15
scikit-learn==1.8.0
scipy==1.17.0
six==1.17.0
sniffio==1.3.1
threadpoolctl==3.6.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
